+++ onmt/translate/PhraseTable.lua
@@ -1,4 +1,3 @@
-
 --[[Parse and lookup a words from a phrase table.
 --]]
 local PhraseTable = torch.class('PhraseTable')
@@ -11,7 +10,8 @@ function PhraseTable:__init(filePath)
 
   for line in f:lines() do
     local c = line:split("|||")
-    self.table[onmt.utils.String.strip(c[1])] = c[2]
+    assert(#c == 2, 'badly formatted phrase table: ' .. line)
+    self.table[onmt.utils.String.strip(c[1])] = onmt.utils.String.strip(c[2])
   end
 
   f:close()
@@ -22,6 +22,7 @@ function PhraseTable:lookup(word)
   return self.table[word]
 end
 
+--[[ Return true if the phrase table contains the source word `word`. ]]
 function PhraseTable:contains(word)
   return self:lookup(word) ~= nil
 end