+++ onmt/translate/Translator.lua
@@ -26,7 +26,7 @@ function Translator:__init(args)
   self.opt = args
   onmt.utils.Cuda.init(self.opt)
 
-  print('Loading \'' .. self.opt.model .. '\'...')
+  _G.logger:info('Loading \'' .. self.opt.model .. '\'...')
   self.checkpoint = torch.load(self.opt.model)
 
   self.models = {}
@@ -276,6 +276,7 @@ function Translator:translate(srcBatch, srcFeaturesBatch, goldBatch, goldFeature
 
     local info = {}
     info.score = predScore[b][1]
+    info.attention = attn[b][1]
     info.nBest = {}
 
     if goldScore ~= nil then
@@ -286,6 +287,7 @@ function Translator:translate(srcBatch, srcFeaturesBatch, goldBatch, goldFeature
       for n = 1, self.opt.n_best do
         info.nBest[n] = {}
         info.nBest[n].tokens = self:buildTargetTokens(pred[b][n], predFeats[b][n], srcBatch[b], attn[b][n])
+        info.nBest[n].attention = attn[b][n]
         info.nBest[n].score = predScore[b][n]
       end
     end