+++ onmt/utils/Parallel.lua
@@ -37,18 +37,22 @@ function Parallel.init(opt)
     Parallel._tds = require('tds')
 
     if Parallel.count > 1 then
-      print('Using ' .. Parallel.count .. ' threads on ' .. #Parallel.gpus .. ' GPUs')
+      _G.logger:info('Using ' .. Parallel.count .. ' threads on ' .. #Parallel.gpus .. ' GPUs')
+      local globalLogger = _G.logger
       local threads = require('threads')
       threads.Threads.serialization('threads.sharedserialize')
       local thegpus = Parallel.gpus
       Parallel._gmutex = threads.Mutex()
       Parallel._pool = threads.Threads(
         Parallel.count,
-        function(threadid)
+        function()
           require('cunn')
           require('nngraph')
           require('onmt.init')
           _G.threads = require('threads')
+        end,
+        function(threadid)
+          _G.logger = globalLogger
           onmt.utils.Cuda.init(opt, thegpus[threadid])
         end
       ) -- dedicate threads to GPUs
@@ -60,10 +64,10 @@ function Parallel.init(opt)
       local ret
       ret, Parallel.usenccl = pcall(require, 'nccl')
       if not ret then
-        print("WARNING: for improved efficiency in nparallel mode - do install nccl")
+        _G.logger:warning("For improved efficiency in nparallel mode - do install nccl")
         Parallel.usenccl = nil
       elseif os.getenv('CUDA_LAUNCH_BLOCKING') == '1' then
-        print("WARNING: CUDA_LAUNCH_BLOCKING set - cannot use nccl")
+        _G.logger:warning("CUDA_LAUNCH_BLOCKING set - cannot use nccl")
         Parallel.usenccl = nil
       end
     end