+++ onmt/utils/Tensor.lua
@@ -21,6 +21,18 @@ local function recursiveClone(out)
   return recursiveApply(out, function (h) return h:clone() end)
 end
 
+--[[ Recursively add `b` tensors into `a`'s. ]]
+local function recursiveAdd(a, b)
+  if torch.isTensor(a) then
+    a:add(b)
+  else
+    for i = 1, #a do
+      recursiveAdd(a[i], b[i])
+    end
+  end
+  return a
+end
+
 local function recursiveSet(dst, src)
   if torch.isTensor(dst) then
     dst:set(src)
@@ -33,15 +45,11 @@ end
 
 --[[ Clone any serializable Torch object. ]]
 local function deepClone(obj)
-  local mem = torch.MemoryFile("w"):binary()
+  local mem = torch.MemoryFile("rw"):binary()
   mem:writeObject(obj)
-
-  local reader = torch.MemoryFile(mem:storage(), "r"):binary()
-  local clone = reader:readObject()
-
-  reader:close()
+  mem:seek(1)
+  local clone = mem:readObject()
   mem:close()
-
   return clone
 end
 
@@ -164,6 +172,7 @@ end
 return {
   recursiveApply = recursiveApply,
   recursiveClone = recursiveClone,
+  recursiveAdd = recursiveAdd,
   recursiveSet = recursiveSet,
   deepClone = deepClone,
   reuseTensor = reuseTensor,