+++ test/test.lua
@@ -274,35 +274,18 @@ end
 
 tester:add(beamSearchTest)
 
-local nmttest = torch.TestSuite()
-
--- local function equal(t1, t2, msg)
---    if (torch.type(t1) == "table") then
---       for k, _ in pairs(t2) do
---          equal(t1[k], t2[k], msg)
---       end
---    else
---       tester:eq(t1, t2, 0.00001, msg)
---    end
--- end
-
-
-function nmttest.Data()
-end
-
-tester:add(nmttest)
-
-function onmt.test(tests, fixedSeed)
+local function main()
   -- Limit number of threads since everything is small
   local nThreads = torch.getnumthreads()
   torch.setnumthreads(1)
 
-   -- Randomize stuff
-  local seed = fixedSeed or (1e5 * torch.tic())
-  print('Seed: ', seed)
-  math.randomseed(seed)
-  torch.manualSeed(seed)
-  tester:run(tests)
+  tester:run()
+
   torch.setnumthreads(nThreads)
-  return tester
+
+  if tester.errors[1] then
+    os.exit(1)
+  end
 end
+
+main()