+++ tools/utils/separators.lua
@@ -1,6 +1,6 @@
 return {
-  sep_marker = '￭',
-  sep_marker_substitute = '■',
+  joiner_marker = '￭',
+  joiner_marker_substitute = '■',
   feat_marker = '￨',
   feat_marker_substitute = '│'
 }